fix(terminal): force dynamic rendering — add use client + force-dynamic [C-271]
